+++ connect/amazon/dynamodb/fake/src/main/kotlin/org/http4k/connect/amazon/dynamodb/grammar/GreaterThan.kt
@@ -6,7 +6,7 @@ import parser4k.ref
 object GreaterThan : ExprFactory {
     override operator fun invoke(parser: () -> Parser<Expr>): Parser<Expr> =
         binaryExpr(ref(parser), ">") { attr1, attr2 ->
-            Expr { item ->
+            expr(attr1, attr2) { item ->
                 item.comparable(attr1) > item.comparable(attr2)
             }
         }